Moving JS file to dedicaded folder
