+++ index.js
@@ -1,7 +1,7 @@
 const express = require('express'),
     app = express(),
-    view_routes = require('./view-routes'),
-    api_routes = require('./api-routes'),
+    view_routes = require('./js/view-routes'),
+    api_routes = require('./js/api-routes'),
     _ = require('lodash'),
     i18next = require('i18next'),
     i18nextMiddleware = require('i18next-express-middleware'),
